MAJOR CHANGE: reordered files to convert to MVC architecture. Changes to *.php files to comply with the reordering. Modified footer to add link to admin login.
