+++ controller/listarMensajes.php
@@ -1,5 +1,8 @@
 <?php
 require "propiedades.php";
+ini_set('display_errors', 1);
+ini_set('display_startup_errors', 1);
+error_reporting(0);
 function listarMensajes($nombre, $apellido, $paisOSha): array
 {
     $datetime = date("Y-m-d H:i:s");
