+++ controller/propiedades.php
