+++ view/contacto.php
@@ -1,22 +1,21 @@
 <!DOCTYPE html>
-<html lang="es">
+<html lang="en">
+
 <head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Contacto</title>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css"
           integrity="sha512-KfkfwYDsLkIlwQp6LFnl8zNdLGxu9YAA1QvwINks4PhcElQSvqcyVLLD9aMhXd13uQjoXtEKNosOWaZqXgel0g=="
-          crossorigin="anonymous" referrerpolicy="no-referrer"
-    />
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet"
-          integrity="sha384-0evHe/X+R7YkIZDRvuzKMRqM+OrBnVFBL6DOitfPri4tjfHxaWutUpFmBp4vmVor" crossorigin="anonymous">
+          crossorigin="anonymous" referrerpolicy="no-referrer">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
+          integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-pprn3073KE6tl6bjs2QrFaJGz5/SUsLqktiwsUTF55Jfv3qYSDhgCecCxMW52nD2"
             crossorigin="anonymous"></script>
     <script defer src="js/buscarSeguimiento.js"></script>
-    <link rel="stylesheet" href="css/cards.css">
     <link rel="stylesheet" href="css/style.css">
     <link rel="stylesheet" href="fonts/stylesheet.css">
-    <script src="js/mostrarPopUp.js" defer></script>
-    <meta http-equiv="Content-Type" content="text/html;charset=UTF-8">
-    <title>Resultados - Tell U</title>
     <script async>(function (w, d) {
             var h = d.head || d.getElementsByTagName("head")[0];
             var s = d.createElement("script");
@@ -25,11 +24,17 @@
             h.appendChild(s);
         })(window, document);</script>
 </head>
-<body>
+
+<body style="background: url('img/img_contacto_2x.png') no-repeat center center fixed;
+-webkit-background-size: cover;
+-moz-background-size: cover;
+background-size: cover;
+-o-background-size: cover;">
+
 <!-- inicio de barra de navegación -->
 <nav class="navbar navbar-expand-lg navbar-light bg-light">
     <div class="container-fluid">
-        <a class="navbar-brand" href="index.php"><img src="img/logo.png" width="50" alt="navigation bar logo"></a>
+        <a class="navbar-brand" href="../index.php"><img src="img/logo.png" width="50" alt="navigation bar logo"></a>
         <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
                 aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
             <span class="navbar-toggler-icon"></span>
@@ -68,58 +73,43 @@
 </nav>
 <!-- fin de barra de navegación -->
 
-<div class="container cardContainer">
-    <h1 class="text-white mb-3">Resultados</h1>
-    <?php
-    include 'php/listarMensajes.php';
-    $registros = array();
+<div class="container text-white">
+    <div class="row">
+        <div class="col text-center">
+            <img class="img-fluid" src="img/logo_blanco.png" alt="">
+            <h1>
+                Conversemos
+            </h1>
 
-    $nombre = ucwords(strtolower($_POST['txtNombre']));
-    $apellido = ucwords(strtolower($_POST['txtApellido']));
-    $pais = ucwords($_POST['cmbPais']);
-    $registros = listarMensajes($nombre, $apellido, $pais);
-    echo "<h2 class=\"text-white mb-5\">Mensajes en <span class=\"searchTerms\">" . $pais . "</span> para <span
-                class=\"searchTerms\">" . $nombre . " " . $apellido . "</span>:</h2>";
-    ?>
-
-    <ol class="messageCard">
-        <?php
-        $index = 1;
-        foreach ($registros as $registro) {
-            $fechaInt = strtotime($registro['fecha']);
-            $fecha = date('d/m/Y', $fechaInt);
-            $nombreOriginal = ucwords(strtolower($registro['nombre']));
-            $apellidoOriginal = ucwords(strtolower($registro['apellido']));
-            echo "<li class=\"messageCardItem\" style=\"--animation-order:" . $index . "\"><a class=\"messageCardLink\">
-                <div class=\"messageCardContent articles__content--lhs\">
-                    <h3 class=\"messageCardTitle\">" . $registro['mensaje'] . "</h3>
-                    <div class=\"messageCardFooter\">
-                        <span>" . $nombreOriginal . " " . $apellidoOriginal . "</span>
-                        <span>" . $fecha . "</span>
-                    </div>
+        </div>
+    </div>
+    <div class="row text-center">
+        <div class="col col-sm-8 col-md-4 mx-auto">
+            <form>
+                <div>
+                    <label for="email" class="form-label">Tu correo: </label> <br>
+                    <INPUT class="form-control" TYPE="email" NAME="email" MAXLENGTH=18 placeholder="correo@example.com">
                 </div>
-                <div class=\"messageCardContent articles__content--rhs\" aria-hidden=\"true\">
-                    <h3 class=\"messageCardTitle\">" . $registro['mensaje'] . "</h3>
-                    <div class=\"messageCardFooter\">
-                        <p>Perú</p>
-                        <span>" . $fecha . "</span>
-                    </div>
+                <div>
+                    <label for="text" class="form-label">Tu comentario</label> <br>
+                    <textarea class="form-control" rows="4" NAME="comentario"
+                              placeholder="Escribe un comentario sobre nosotros..."></textarea>
                 </div>
-            </a>
-        </li>";
-            $index++;
-        }
-        ?>
-    </ol>
+                <div class="py-5">
+                    <button type="submit" class="boton-efecto">Enviar</button>
+                </div>
+            </form>
+        </div>
+    </div>
 </div>
-
 <!-- inicio de footer -->
-<footer class="footer mt-auto py-3 bg-light text-center fixed-bottom ">
-    <div class="container ">
-        <span class="text-muted ">Página desarrollada por el grupo 4</span>
+<footer class="footer mt-auto py-3 bg-light text-center fixed-bottom">
+    <div class="container">
+        <span class="text-muted ">Página desarrollada por el grupo 4.  <a class="text-muted" href="adminlogin.php">Admin login</a></span>
     </div>
 </footer>
 <!-- fin de footer -->
 
 </body>
+
 </html>
