+++ view/css/sentmessage.css
